requirements.txt
numpy==1.21.1
matplotlib==3.4.2
adjustText==0.7.3
pandas==1.3.1
tqdm
pyyaml